+++ test/plugins/template.engine/cases/case46.js
@@ -1,15 +1,17 @@
-export const title = '1) load a simple SVG';
-export const description = `load an external SVG`;
+export const title = '46) g-content with $$.fromURL() and $$.currentContent()';
+export const description = `load an external SVG with $$.svg()`;
 
 export async function script () {
   const svg = gSVG(document.querySelector('#svg'));
-  await svg.render();
+  svg.render();
   svg.querySelector('g').addEventListener('load', () => {
     document.querySelector('#result').innerHTML = sourceFormat(svg.source());
   });
 }
 
 export default `<svg viewBox="0 0 100 100" id="svg" style="width: 100px; height: 100px">
-  <g g-content="$$.url('/test/plugins/load/assets/image.svg')"></g>
+  <g g-content="$$.fromURL($$.currentContent().trim())">
+    /test/plugins/template.engine/assets/image.svg
+  </g>
 </svg>
 <pre id="result"></pre>`;