+++ test/plugins/template.engine/cases/case48.js
@@ -1,4 +1,4 @@
-export const title = '4) g-load and g-for';
+export const title = '48) g-load with $$.fromURL() in g-for';
 export const description = `load an external SVG in a g-for`;
 
 export function script () {
@@ -12,7 +12,8 @@ export function script () {
 export default `<svg viewBox="0 0 500 50" id="svg" style="width: 500px; height: 100px"">
   <defs g-for="n of 10">
     <g g-bind:transform="$$.translate(n * 50, 0)" 
-       g-load="$$.url('/test/plugins/load/assets/image.svg').width(50).height(50)"></g>
+       g-content="$$.fromURL('/test/plugins/template.engine/assets/image.svg')"
+       g-on:load="event => event.detail.querySelector('svg').width(50).height(50)"></g>
   </defs>
 </svg>
 <pre id="result"></pre>`;