+++ test/plugins/template.engine/cases/case50.js
@@ -1,5 +1,5 @@
-export const title = '7) g-content';
-export const description = `previous behaviour`;
+export const title = '50) g-load with $$.fromURL() and text file';
+export const description = `load an external TXT file`;
 
 export async function script () {
   const svg = gSVG(document.querySelector('#svg'));
@@ -10,6 +10,6 @@ export async function script () {
 }
 
 export default `<svg viewBox="0 0 500 50" id="svg" style="width: 500px; height: 100px">
-  <text x="0" y="0" g-content="'/test/plugins/load/assets/text.txt'">hola, mundo</text>
+  <text x="0" y="10" g-content="$$.fromURL('/test/plugins/template.engine/assets/text.txt')">...</text>
 </svg>
 <pre id="result"></pre>`;